refactor(claude-skills): split update-bsg-skills.py into bsg_updater package
The 511 LOC update-bsg-skills.py exceeded the 500 LOC oversized
threshold surfaced by the tech-lead audit. Split responsibilities into
a bsg_updater/ package (config, log_setup, http_client, manifest,
installer, walker, reconcile, settings, core), each well under 500 LOC.

The public entry point stays a single file at
claude-skills/scripts/update-bsg-skills.py so the INSTALL.md flow
("drop one Python script and run it") is preserved. On first install
the entry point bootstraps the helper package from
raw.githubusercontent.com; subsequent runs reuse the on-disk package,
which the manifest reconcile keeps refreshed alongside every other BSG
file.

Tests updated to import bsg_updater submodules directly and patch at
the boundary they actually use.

Closes #692
